requirements.txt
numpy
matplotlib
networkx
jupyter
pyyaml
